+++ implementation/src/draw/gui.cpp
@@ -31,9 +31,9 @@ void initImGuiWindows() {
   // settings window
   imguiwindow::SHOW_SETTINGS_WINDOW = imguiwindow::INITIAL_SHOW_SETTINGS_WINDOW;
   for (const ProblemType& type : problemType::PROBLEM_TYPES) {
-    imguiwindow::ACTIVE[type]    = imguiwindow::INITIAL_ACTIVENESS.at(type);
-    imguiwindow::COLOR[type]     = imguiwindow::INITIAL_COLOR.at(type);
-    imguiwindow::THICKNESS[type] = imguiwindow::INITIAL_THICKNESS.at(type);
+    imguiwindow::ACTIVE[(unsigned int) type]    = imguiwindow::INITIAL_ACTIVENESS.at((unsigned int) type);
+    imguiwindow::COLOR[(unsigned int) type]     = imguiwindow::INITIAL_COLOR.at((unsigned int) type);
+    imguiwindow::THICKNESS[(unsigned int) type] = imguiwindow::INITIAL_THICKNESS.at((unsigned int) type);
   }
   imguiwindow::VERTEX_COLOR = imguiwindow::INITIAL_VERTEX_COLOR;
 }
@@ -51,13 +51,15 @@ void drawImgui() {
       "Application average %.3f ms/frame (%.1f FPS)", 1000.0f / ImGui::GetIO().Framerate, ImGui::GetIO().Framerate);
 
     ImGui::Spacing();
-    ImGui::Checkbox("BTSP exact", &imguiwindow::ACTIVE[ProblemType::BTSP_exact]);
-    ImGui::SliderFloat("thickness##BTSP exact", &imguiwindow::THICKNESS[ProblemType::BTSP_exact], 0.0f, 20.0f, "%.1f");
-    ImGui::ColorEdit3("##BTSP exact", (float*) &imguiwindow::COLOR[ProblemType::BTSP_exact]);
+    ImGui::Checkbox("BTSP exact", &imguiwindow::ACTIVE[(unsigned int) ProblemType::BTSP_exact]);
+    ImGui::SliderFloat(
+      "thickness##BTSP exact", &imguiwindow::THICKNESS[(unsigned int) ProblemType::BTSP_exact], 0.0f, 20.0f, "%.1f");
+    ImGui::ColorEdit3("##BTSP exact", (float*) &imguiwindow::COLOR[(unsigned int) ProblemType::BTSP_exact]);
 
-    ImGui::Checkbox("TSP  exact", &imguiwindow::ACTIVE[ProblemType::TSP_exact]);
-    ImGui::ColorEdit3("##TSP exact", (float*) &imguiwindow::COLOR[ProblemType::TSP_exact]);
-    ImGui::SliderFloat("thickness##TSP exact", &imguiwindow::THICKNESS[ProblemType::TSP_exact], 0.0f, 30.0f);
+    ImGui::Checkbox("TSP  exact", &imguiwindow::ACTIVE[(unsigned int) ProblemType::TSP_exact]);
+    ImGui::ColorEdit3("##TSP exact", (float*) &imguiwindow::COLOR[(unsigned int) ProblemType::TSP_exact]);
+    ImGui::SliderFloat(
+      "thickness##TSP exact", &imguiwindow::THICKNESS[(unsigned int) ProblemType::TSP_exact], 0.0f, 30.0f);
 
     ImGui::ColorEdit3("vertex color", (float*) &imguiwindow::VERTEX_COLOR);
     ImGui::End();
