+++ implementation/src/main.cpp
@@ -84,9 +84,10 @@ int main([[maybe_unused]] int argc, [[maybe_unused]] char* argv[]) {
 
   Buffers buffers;
   buffers.coordinates.bind();
-  buffers.coordinates.bufferData(graph::POINTS_F, 2);              // components per vertex
-  buffers.tourCoordinates.bufferData(graph::POINTS_F);             // copy vertex coordinates also into shader buffer
-  buffers.tour.bufferData(graph::ORDER[ProblemType::BTSP_exact]);  // copy indices of verteces in tour to shader buffer
+  buffers.coordinates.bufferData(graph::POINTS_F, 2);   // components per vertex
+  buffers.tourCoordinates.bufferData(graph::POINTS_F);  // copy vertex coordinates also into shader buffer
+  buffers.tour.bufferData(
+    graph::ORDER[(unsigned int) ProblemType::BTSP_exact]);  // copy indices of verteces in tour to shader buffer
 
   VertexArray vao;
   vao.bind();
